Move common types and constants into a separate file
